swap, types: move Uint256 implementation to new swarm/types package
