+++ swap/common_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/abi/bind/backends"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
-	"github.com/ethereum/go-ethereum/core/types"
+	etypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/p2p"
 	"github.com/ethereum/go-ethereum/p2p/simulations/adapters"
@@ -24,6 +24,7 @@ import (
 	"github.com/ethersphere/swarm/network"
 	"github.com/ethersphere/swarm/p2p/protocols"
 	"github.com/ethersphere/swarm/state"
+	"github.com/ethersphere/swarm/types"
 )
 
 // swapTestBackend encapsulates the SimulatedBackend and can offer
@@ -167,7 +168,7 @@ func newTestCheque() *Cheque {
 	cheque := &Cheque{
 		ChequeParams: ChequeParams{
 			Contract:         testChequeContract,
-			CumulativePayout: Uint64ToUint256(42),
+			CumulativePayout: types.Uint64ToUint256(42),
 			Beneficiary:      beneficiaryAddress,
 		},
 		Honey: uint64(42),
@@ -180,7 +181,7 @@ func newSignedTestCheque(testChequeContract common.Address, beneficiaryAddress c
 	cheque := &Cheque{
 		ChequeParams: ChequeParams{
 			Contract:         testChequeContract,
-			CumulativePayout: Uint64ToUint256(cumulativePayout.Uint64()),
+			CumulativePayout: types.Uint64ToUint256(cumulativePayout.Uint64()),
 			Beneficiary:      beneficiaryAddress,
 		},
 		Honey: cumulativePayout.Uint64(),
@@ -201,7 +202,7 @@ func newRandomTestCheque() *Cheque {
 	cheque := &Cheque{
 		ChequeParams: ChequeParams{
 			Contract:         testChequeContract,
-			CumulativePayout: Uint64ToUint256(amount),
+			CumulativePayout: types.Uint64ToUint256(amount),
 			Beneficiary:      beneficiaryAddress,
 		},
 		Honey: amount,
@@ -224,7 +225,7 @@ func testCashCheque(s *Swap, cheque *Cheque) {
 }
 
 // when testing, we don't need to wait for a transaction to be mined
-func testWaitForTx(ctx context.Context, backend cswap.Backend, tx *types.Transaction) (*types.Receipt, error) {
+func testWaitForTx(ctx context.Context, backend cswap.Backend, tx *etypes.Transaction) (*etypes.Receipt, error) {
 
 	var stb *swapTestBackend
 	var ok bool
@@ -237,7 +238,7 @@ func testWaitForTx(ctx context.Context, backend cswap.Backend, tx *types.Transac
 	if err != nil {
 		return nil, err
 	}
-	if receipt.Status != types.ReceiptStatusSuccessful {
+	if receipt.Status != etypes.ReceiptStatusSuccessful {
 		return nil, cswap.ErrTransactionReverted
 	}
 	return receipt, nil
